requirements.txt
streamlit>=1.37.0
pymupdf>=1.24.9
python-pptx>=0.6.23
chromadb>=0.5.3
llama-index>=0.10.50
llama-index-vector-stores-chroma>=0.1.11
llama-index-embeddings-huggingface>=0.1.4
llama-index-llms-ollama>=0.1.3
llama-index-retrievers-bm25>=0.1.3


gradio>=6.1.0
gradio-pdf>=0.0.24
mineru-vl-utils[transformers]>=0.1.17.1
pdf2image>=1.17.0
pillow>=12.0.0
pypdf>=6.4.0
pypdfium2>=5.2.0
reportlab>=4.4.6
loguru>=0.7.3
